requirements.txt
numpy
viam-sdk==0.25.2
scikit-learn
scipy
matplotlib
pylint
pytest
pytest-asyncio
